管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -32,6 +32,8 @@ public class EmployeeAction extends ActionBase {
 
     public void index() throws ServletException, IOException {
 
+        if(checkAdmin()) {
+
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
@@ -50,19 +52,25 @@ public class EmployeeAction extends ActionBase {
 
         forward(ForwardConst.FW_EMP_INDEX);
 
+        }
+
     }
 
     public void entryNew() throws ServletException, IOException {
+
+        if(checkAdmin()) {
 
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
         forward(ForwardConst.FW_EMP_NEW);
+
+        }
     }
 
     public void create() throws ServletException, IOException{
 
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
 
             EmployeeView ev = new EmployeeView(
                     null,
@@ -97,6 +105,8 @@ public class EmployeeAction extends ActionBase {
 
     public void show() throws ServletException, IOException{
 
+        if(checkAdmin()) {
+
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -108,9 +118,13 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_SHOW);
+
+        }
     }
 
     public void edit() throws ServletException, IOException{
+
+        if(checkAdmin()) {
 
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -124,11 +138,13 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_EDIT);
+
+        }
     }
 
     public void update() throws ServletException, IOException{
 
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
 
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -161,11 +177,27 @@ public class EmployeeAction extends ActionBase {
 
     public void destroy() throws ServletException, IOException{
 
-        if(checkToken()){
+        if(checkAdmin() && checkToken()){
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
             putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
+    }
+
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
         }
+
+    }
 
 }
